requirements.txt
python-socketio[client]
pyserial
python-dotenv
RPi.GPIO
